requirements.txt
bluecat-libraries
progressbar2
psycopg2
python-dotenv
requests
setuptools
